// src/routes/Product/enums/product.enums.ts
export enum IdealFor {
  MEN = 'Men',
  WOMEN = 'Women',
}

export enum SleeveLength {
  THREE_QUARTER = '3/4 Sleeve',
  FULL = 'Full Sleeve',
  HALF = 'Half Sleeve',
  NA = 'NA',
  ROLL_UP = 'Roll-up Sleeve',
  SHORT = 'Short Sleeve',
  SLEEVELESS = 'Sleeveless',
}

export enum Pattern {
  ABSTRACT = 'Abstract',
  ANIMAL_PRINT = 'Animal Print',
  APPLIQUE = 'Applique',
  BANDHANI = 'Bandhani',
  CHECKERED = 'Checkered',
  CHIKANKARI_EMBROIDERY = 'Chikankari Embroidery',
  COLORBLOCK = 'Colorblock',
  DYED = 'Dyed',
  EMBELLISHED = 'Embellished',
  EMBROIDERED = 'Embroidered',
  FLORAL_PRINT = 'Floral Print',
  GEOMETRIC_PRINT = 'Geometric Print',
  LEHERIYA = 'Leheriya',
  PAISLEY = 'Paisley',
  POLKA_PRINT = 'Polka Print',
  PRINTED = 'Printed',
  SELF_DESIGN = 'Self Design',
  SOLID = 'Solid',
  STRIPED = 'Striped',
  TRIBAL_PRINT = 'Tribal Print',
  WASHED = 'Washed',
  WOVEN = 'Woven',
}

export enum Fabric {
  ACRYLIC_BLEND = 'Acrylic Blend',
  ART_SILK = 'Art Silk',
  CHANDERI = 'Chanderi',
  CHIFFON = 'Chiffon',
  COTTON_BLEND = 'Cotton Blend',
  COTTON_RAYON = 'Cotton Rayon',
  COTTON_SILK = 'Cotton Silk',
  CREPE = 'Crepe',
  DUPION_SILK = 'Dupion Silk',
  GEORGETTE = 'Georgette',
  JACQUARD = 'Jacquard',
  JUTE = 'Jute',
  KHADI_COTTON = 'Khadi Cotton',
  KHADI_SILK = 'Khadi Silk',
  LACE = 'Lace',
  LINEN = 'Linen',
  LINEN_BLEND = 'Linen Blend',
  MODAL = 'Modal',
  MUSLIN = 'Muslin',
  NET = 'Net',
  POLYESTER = 'Polyester',
  PURE_COTTON = 'Pure Cotton',
  PURE_LINEN = 'Pure Linen',
  PURE_SILK = 'Pure Silk',
  SATIN = 'Satin',
  SILK_BLEND = 'Silk Blend',
  TUSSAR_SILK = 'Tussar Silk',
  VELVET = 'Velvet',
  VISCOSE_RAYON = 'Viscose Rayon',
  WOOL = 'Wool',
}

export enum TopType {
  // Men
  KURTA = 'Kurta',
  KURTI = 'Kurti',
  NA = 'NA',
  SHERWANI = 'Sherwani',
  SHIRT = 'Shirt',
  // Women-only
  CROP_TOP = 'Crop Top',
  ETHNIC_TOP = 'Ethnic Top',
  KAFTAN = 'Kaftan',
}

export enum BottomType {
  // Men
  DHOTI = 'Dhoti',
  DHOTI_PANT = 'Dhoti Pant',
  NA = 'NA',
  PANT = 'Pant',
  PYJAMA = 'Pyjama',
  SALWAR = 'Salwar',
  // Women-only
  CHURIDAR = 'Churidar',
  PALAZZO = 'Palazzo',
  PATIALA = 'Patiala',
  SARI_PANT = 'Sari Pant',
  SHARARA = 'Sharara',
  SKIRT = 'Skirt',
}

export enum AdditionalGarments {
  ATTACHED_DUPATTA = 'Attached Dupatta',
  CAPE = 'Cape',
  DUPATTA = 'Dupatta',
  ETHNIC_JACKET = 'Ethnic Jacket',
  NA = 'NA',
}

export enum Size {
  S0 = '0', S1 = '1', S2 = '2', S3 = '3', S4 = '4', S5 = '5', S6 = '6',
  S8 = '8', S10 = '10', S12 = '12', S14 = '14', S16 = '16', S18 = '18',
  S20 = '20', S22 = '22', S24 = '24', S25 = '25', S26 = '26', S27 = '27',
  S28 = '28', S30 = '30', S32 = '32', S34 = '34', S36 = '36', S37 = '37',
  S38 = '38', S40 = '40', S42 = '42', S44 = '44', S46 = '46', S48 = '48',
  S50 = '50', S52 = '52', S54 = '54',
  XS = 'XS', XXS = 'XXS', S3XS = '3XS',
  S = 'S', M = 'M', L = 'L', XL = 'XL', XXL = 'XXL',
  S3XL = '3XL', S4XL = '4XL', S5XL = '5XL', S6XL = '6XL',
  S7XL = '7XL', S8XL = '8XL', S9XL = '9XL', S10XL = '10XL',
  ONE_SIZE = 'One Size',
}

export enum Color {
  BEIGE = 'Beige',
  BLACK = 'Black',
  BLUE = 'Blue',
  BROWN = 'Brown',
  DARK_BLUE = 'Dark Blue',
  DARK_GREEN = 'Dark Green',
  GOLD = 'Gold',
  GREEN = 'Green',
  GREY = 'Grey',
  LIGHT_BLUE = 'Light Blue',
  LIGHT_GREEN = 'Light Green',
  MAROON = 'Maroon',
  MULTICOLOR = 'Multicolor',
  ORANGE = 'Orange',
  PINK = 'Pink',
  PURPLE = 'Purple',
  RED = 'Red',
  SILVER = 'Silver',
  WHITE = 'White',
  YELLOW = 'Yellow',
}

export enum Occasion {
  CASUAL = 'Casual',
  FESTIVE_PARTY = 'Festive & Party',
  FORMAL = 'Formal',
  WEDDING = 'Wedding',
}

export enum TopFabric {
  ART_SILK = 'Art Silk',
  COTTON = 'Cotton',
  COTTON_BLEND = 'Cotton Blend',
  CREPE = 'Crepe',
  GEORGETTE = 'Georgette',
  MODAL = 'Modal',
  NET = 'Net',
  POLY_CREPE = 'Poly Crepe',
  POLYESTER = 'Polyester',
  RAYON = 'Rayon',
  SILK_BLEND = 'Silk Blend',
  VELVET = 'Velvet',
  VISCOSE = 'Viscose',
}

export enum BottomFabric {
  CHIFFON = 'Chiffon',
  CORDUROY = 'Corduroy',
  COTTON = 'Cotton',
  CREPE = 'Crepe',
  GEORGETTE = 'Georgette',
  KHADI = 'Khadi',
  LACE = 'Lace',
  NYLON = 'Nylon',
  POLYESTER = 'Polyester',
  RAYON = 'Rayon',
  SATIN = 'Satin',
  SILK = 'Silk',
  SYNTHETIC = 'Synthetic',
  VELVET = 'Velvet',
  VISCOSE = 'Viscose',
}

export enum KurtaStyleType {
  A_LINE = 'A-line',
  ANARKALI = 'Anarkali',
  ANGARKHA = 'Angarkha',
  ASYMMETRIC = 'Asymmetric',
  FLARED = 'Flared',
  STRAIGHT = 'Straight',
}

export enum OrnamentationType {
  AARI_WORK = 'Aari Work',
  BEADS_AND_STONES = 'Beads and Stones',
  BUTTON_LOOP = 'Button Loop',
  BUTTONS = 'Buttons',
  CHIKANKARI = 'Chikankari',
  CUT_WORK = 'Cut Work',
  EPAULETTES = 'Epaulettes',
  GOTTA_PATTI = 'Gotta Patti',
  KANTHA_WORK = 'Kantha Work',
  LACE = 'Lace',
  METAL_EMBELLISHMENT = 'Metal Embellishment',
  MIRROR_WORK = 'Mirror Work',
  MUKAISH = 'Mukaish',
  NONE = 'None',
  PATCH_WORK = 'Patch Work',
  PEARL = 'Pearl',
  PHULKARI = 'Phulkari',
  PIPING_DETAIL = 'Piping Detail',
  POCKET_FLAP = 'Pocket Flap',
  POM_POMS = 'Pom Poms',
  SCHIFFILY = 'Schiffily',
  SEQUINS = 'Sequins',
  SHIMMER = 'Shimmer',
  TASSELS = 'Tassels',
  THREAD_WORK = 'Thread Work',
  ZARDOSI = 'Zardosi',
  ZARDOZI = 'Zardozi',
  ZARI = 'Zari',
}

export enum Neck {
  ASYMMETRIC_NECK = 'Asymmetric Neck',
  BOAT_NECK = 'Boat Neck',
  CHOKER_NECK = 'Choker Neck',
  COLLARED_NECK = 'Collared Neck',
  COWL_NECK = 'Cowl Neck',
  HALTER_NECK = 'Halter Neck',
  HENLEY_NECK = 'Henley Neck',
  HIGH_NECK = 'High Neck',
  KEY_HOLE_NECK = 'Key Hole Neck',
  LAPEL_COLLAR = 'Lapel Collar',
  MANDARIN_COLLAR = 'Mandarin Collar',
  NA = 'NA',
  OFF_SHOULDER = 'Off Shoulder',
  ONE_SHOULDER_NECK = 'One Shoulder Neck',
  PETER_PAN_COLLAR = 'Peter Pan Collar',
  ROUND_NECK = 'Round Neck',
  RUFFLED_NECK = 'Ruffled Neck',
  SCOOP_NECK = 'Scoop Neck',
  SHAWL_COLLAR = 'Shawl Collar',
  SQUARE_NECK = 'Square Neck',
  SWEETHEART_NECK = 'Sweetheart Neck',
  TIE_UP = 'Tie-up',
  U_NECK = 'U - Neck',
  V_NECK = 'V Neck',
}

export enum Trend {
  AJRAKH = 'Ajrakh',
  ANARKALIS = 'Anarkalis',
  ARCHITECTURAL_PRINT = 'Architectural Print',
  BANDHEJ = 'Bandhej',
  BLOCK_PRINT = 'Block print',
  CHIKANKARI = 'Chikankari',
  CLASSIC = 'Classic',
  COLORBLOCK = 'Colorblock',
  GEOMETRIC_PRINT = 'Geometric Print',
  IKAT = 'Ikat',
  IKKAT = 'Ikkat',
  INDIGO = 'Indigo',
  KURTA_PALAZZO_SETS = 'Kurta Palazzo Sets',
  KURTA_SKIRT_SET = 'Kurta-skirt set',
  MIRROR_WORK = 'Mirror work',
  PASTELS = 'Pastels',
  PINTUCK_DETAILS = 'Pintuck Details',
  PLACEMENT_EMBROIDERIES = 'Placement Embroideries',
  PLACEMENT_PRINT = 'Placement Print',
  PRINTED_BORDER = 'Printed Border',
  SEQUENCED = 'Sequenced',
  SHIBORI = 'Shibori',
  STYLISH_STRIPES = 'Stylish Stripes',
  TASSELS = 'Tassels',
  TIE_DYE = 'Tie dye',
  TRADITIONAL_SHERWANI = 'Traditional Sherwani',
  TRADITIONAL_COLLECTION = 'Traditional collection',
}

export enum PatternPrintType {
  ABSTRACT = 'Abstract',
  AJRAKH = 'Ajrakh',
  ANIMAL_PRINT = 'Animal Print',
  APPLIQUE = 'Applique',
  ARGYLE = 'Argyle',
  AZTEC = 'Aztec',
  BANDHNI = 'Bandhni',
  BATIK = 'Batik',
  BLOCK_PRINT = 'Block Print',
  BOTANICAL_PRINT = 'Botanical Print',
  CAMOUFLAGE = 'Camouflage',
  CHECKERED = 'Checkered',
  CHEVRON_ZIG_ZAG = 'Chevron/Zig Zag',
  COLORBLOCK = 'Colorblock',
  CONVERSATIONAL = 'Conversational',
  DIGITAL_PRINT = 'Digital Print',
  DITSY = 'Ditsy',
  EMBELLISHED = 'Embellished',
  EMBROIDERED = 'Embroidered',
  ETHNIC_MOTIFS = 'Ethnic Motifs',
  FLORAL = 'Floral',
  FOIL_PRINT = 'Foil Print',
  GEOMETRIC = 'Geometric',
  GOLD_PRINT = 'Gold Print',
  GRAPHIC = 'Graphic',
  HERRINGBONE = 'Herringbone',
  HOUNDSTOOTH = 'Houndstooth',
  HUMAN_FIGURES = 'Human Figures',
  IKKAT = 'Ikkat',
  KALAMKARI = 'Kalamkari',
  LEHERIYA = 'Leheriya',
  LIGHT_WASH = 'Light wash',
  OMBRE = 'Ombre',
  PAISLEY = 'Paisley',
  POLKA = 'Polka',
  QUIRKY = 'Quirky',
  SCHIFFILY = 'Schiffily',
  SELF_DESIGN = 'Self Design',
  SEQUINED = 'Sequined',
  SHIBORI = 'Shibori',
  SOLID = 'Solid',
  STRIPED = 'Striped',
  TIE_DYE = 'Tie & Dye',
  TRIBAL = 'Tribal',
  TYPOGRAPHY = 'Typography',
  VARSITY = 'Varsity',
  WARLI = 'Warli',
  WOVEN = 'Woven',
}

export enum SleeveStyle {
  BATWING = 'Batwing Sleeve',
  BELL = 'Bell Sleeve',
  BISHOP = 'Bishop Sleeve',
  CAP = 'Cap Sleeve',
  CAPE = 'Cape Sleeves',
  COLD_SHOULDER = 'Cold Shoulder Sleeve',
  CUFFED = 'Cuffed Sleeve',
  CUTOUT = 'Cutout',
  DOLMAN = 'Dolman Sleeve',
  EXTENDED = 'Extended Sleeve',
  FLARED = 'Flared Sleeve',
  FLUTE = 'Flute Sleeve',
  FLUTTER = 'Flutter Sleeve',
  KAFTAN = 'Kaftan Sleeve',
  KIMONO = 'Kimono Sleeve',
  LAYERED = 'Layered',
  NA = 'NA',
  NO_SLEEVE = 'No Sleeve',
  PETAL = 'Petal Sleeve',
  PUFF = 'Puff Sleeve',
  RAGLAN = 'Raglan Sleeve',
  REGULAR = 'Regular Sleeve',
  ROLL_UP = 'Roll Up Sleeve',
  RUFFLED = 'Ruffled Sleeve',
  SHOULDER_STRAPS = 'Shoulder Straps',
  SLIT = 'Slit Sleeve',
  TIE_UP = 'Tie-up',
}

export enum DetailPlacement {
  ALL_OVER = 'All - Over',
  BACK = 'Back',
  CUFF = 'Cuff',
  FRONT_PANEL = 'Front Panel',
  HEMLINE = 'Hemline',
  NECKLINE = 'Neckline',
  NONE = 'None',
  POCKET = 'Pocket',
  SIDE = 'Side',
  SLEEVE = 'Sleeve',
  SLITS = 'Slits',
  WAIST = 'Waist',
  YOKE = 'Yoke',
}

export enum SurfaceStyling {
  APPLIQUE = 'Applique',
  BOW = 'Bow',
  CROCHETS_SCHIFFILY = 'Crochets & Schiffily',
  CUT_OUTS = 'Cut-Outs',
  FRINGED = 'Fringed',
  KNOTS = 'Knots',
  LACES = 'Laces',
  NONE = 'None',
  POCKETS = 'Pockets',
  RUFFLE = 'Ruffle',
  TIE_UPS = 'Tie-Ups',
}

export enum TopLength {
  ABOVE_KNEE = 'Above Knee',
  ANKLE_LENGTH = 'Ankle Length',
  BELOW_KNEE = 'Below Knee',
  CALF_LENGTH = 'Calf Length',
  CROP_LENGTH = 'Crop Length',
  FLOOR_LENGTH = 'Floor Length',
  HIP_LENGTH = 'Hip Length',
  KNEE_LENGTH = 'Knee Length',
  NA = 'NA',
  THIGH_LENGTH = 'Thigh Length',
  WAIST_LENGTH = 'Waist Length',
}

export enum ListingStatus {
  ACTIVE = 'Active',
  INACTIVE = 'Inactive',
}